requirements.txt
SQLAlchemy
PyYAML
matplotlib
requests
beautifulsoup4
moviepy
pillow
WTForms
Flask
Flask-Login